JNJ-2450 Prevent update creating empty gene list
Currently, we do not know allow an individual to create an empty gene
list. However, a user can delete all of the genes from a gene list so we
should not allow this behavior as it is inconsistent with the creation
behavior.

+++ grails-app/controllers/GeneSignatureController.groovy
@@ -547,6 +547,13 @@ class GeneSignatureController {
 	    	delParam.each { delItems.add(it) }
 	    }
 
+		int j = gs.geneSigItems.size()
+		
+		if(delItems.size() == j ) {
+			flash.message = "<div class='warning'>You tried to delete all items. You must leave at least one item</div>"
+			return render(view: 'edit_items', model:[gs:gs, errorFlag:true])
+		}
+		
 		// delete indicated ids
 		gs = geneSignatureService.deleteGeneSigItems(gs, delItems)
